- Add DetailViewController showing webview of user detail - Paging handling

+++ SenditProject/ItemModel.swift
@@ -9,21 +9,11 @@
 import Foundation
 
 struct ItemModel {
-    var id: Int16!
+    var id: Int!
     var name: String!
     var desc: String!
     var icon: String!
     var timestamp: Int64!
     var url: String!
     var image: String!
-
-//    init(aData: [String: Any]) {
-//        self.id = aData["id"] as! Int16
-//        self.name = aData["name"] as! String
-//        self.desc = aData["description"] as! String
-//        self.icon = aData["icon"] as! String
-//        self.timestamp = aData["timestamp"] as! Int64
-//        self.url = aData["url"] as! String
-//        self.image = aData["image"] as! String
-//    }
 }
